requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile dev.in
#
alabaster==0.7.13
    # via sphinx
babel==2.14.0
    # via sphinx
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   pip-tools
    #   typer
docutils==0.20.1
    # via
    #   myst-parser
    #   sphinx
    #   sphinx-rtd-theme
idna==3.6
    # via requests
imagesize==1.4.1
    # via sphinx
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via
    #   -r dev.in
    #   myst-parser
    #   sphinx
markdown-it-py==3.0.0
    # via
    #   mdit-py-plugins
    #   myst-parser
markupsafe==2.1.3
    # via jinja2
mdit-py-plugins==0.4.0
    # via myst-parser
mdurl==0.1.2
    # via markdown-it-py
myst-parser==2.0.0
    # via -r dev.in
packaging==23.2
    # via
    #   build
    #   pytest
    #   sphinx
pip-tools==7.3.0
    # via -r dev.in
pluggy==1.3.0
    # via pytest
pygments==2.17.2
    # via sphinx
pyproject-hooks==1.0.0
    # via build
pytest==7.4.3
    # via
    #   -r dev.in
    #   pytest-unordered
pytest-unordered==0.5.2
    # via -r dev.in
python-frontmatter==1.0.1
    # via -r dev.in
pyyaml==6.0.1
    # via
    #   myst-parser
    #   python-frontmatter
requests==2.31.0
    # via sphinx
snowballstemmer==2.2.0
    # via sphinx
sphinx==7.2.6
    # via
    #   -r dev.in
    #   myst-parser
    #   sphinx-rtd-theme
    #   sphinxcontrib-applehelp
    #   sphinxcontrib-devhelp
    #   sphinxcontrib-htmlhelp
    #   sphinxcontrib-jquery
    #   sphinxcontrib-qthelp
    #   sphinxcontrib-serializinghtml
sphinx-rtd-theme==2.0.0
    # via -r dev.in
sphinxcontrib-applehelp==1.0.7
    # via sphinx
sphinxcontrib-devhelp==1.0.5
    # via sphinx
sphinxcontrib-htmlhelp==2.0.4
    # via sphinx
sphinxcontrib-jquery==4.1
    # via sphinx-rtd-theme
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.6
    # via sphinx
sphinxcontrib-serializinghtml==1.1.9
    # via sphinx
typer==0.9.0
    # via -r dev.in
typing-extensions==4.9.0
    # via typer
urllib3==2.1.0
    # via requests
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
